Moved code to cpp file from header file
