+++ src/include/optimizer/statistics/selectivity_util.h
@@ -157,79 +157,4 @@ class SelectivityUtil {
     return DEFAULT_SELECTIVITY;
   }
 };
-
-template <typename T>
-double SelectivityUtil::ComputeSelectivity(common::ManagedPointer<NewColumnStats<T>> column_stats,
-                                           const ValueCondition &condition) {
-  switch (condition.GetType()) {
-    case parser::ExpressionType::COMPARE_LESS_THAN:
-      return LessThan(column_stats, condition);
-    case parser::ExpressionType::COMPARE_GREATER_THAN:
-      return GreaterThan(column_stats, condition);
-    case parser::ExpressionType::COMPARE_LESS_THAN_OR_EQUAL_TO:
-      return LessThanOrEqualTo(column_stats, condition);
-    case parser::ExpressionType::COMPARE_GREATER_THAN_OR_EQUAL_TO:
-      return GreaterThanOrEqualTo(column_stats, condition);
-    case parser::ExpressionType::COMPARE_EQUAL:
-      return Equal(column_stats, condition);
-    case parser::ExpressionType::COMPARE_NOT_EQUAL:
-      return NotEqual(column_stats, condition);
-    case parser::ExpressionType::COMPARE_LIKE:
-      return Like(column_stats, condition);
-    case parser::ExpressionType::COMPARE_NOT_LIKE:
-      return NotLike(column_stats, condition);
-    case parser::ExpressionType::COMPARE_IN:
-      return In(column_stats, condition);
-    case parser::ExpressionType::COMPARE_IS_DISTINCT_FROM:
-      return DistinctFrom(column_stats, condition);
-    default:
-      OPTIMIZER_LOG_WARN("Expression type {0} not supported for computing selectivity",
-                         ExpressionTypeToString(condition.GetType(), false).c_str());
-      return DEFAULT_SELECTIVITY;
-  }
-}
-
-template <typename T>
-double SelectivityUtil::LessThanOrEqualTo(common::ManagedPointer<NewColumnStats<T>> column_stats,
-                                          const ValueCondition &condition) {
-  const auto value = condition.GetPointerToValue()->Peek<decltype(T::val_)>();
-
-  // Return default selectivity if empty column_stats
-  if (column_stats == nullptr) {
-    return DEFAULT_SELECTIVITY;
-  }
-
-  // Use histogram to estimate selectivity
-  auto histogram = column_stats->GetHistogram();
-
-  if (value <= histogram->GetMinValue()) return 0;
-  if (value > histogram->GetMaxValue()) return 1;
-  double res =
-      static_cast<double>(histogram->EstimateItemCount(value)) / static_cast<double>(column_stats->GetNumRows());
-  NOISEPAGE_ASSERT(res >= 0 && res <= 1, "res must be within valid range");
-  return res;
-}
-
-template <typename T>
-double SelectivityUtil::Equal(common::ManagedPointer<NewColumnStats<T>> column_stats, const ValueCondition &condition) {
-  // Convert value type to raw value (double)
-  const auto value = condition.GetPointerToValue()->Peek<decltype(T::val_)>();
-  if (column_stats == nullptr) {
-    OPTIMIZER_LOG_DEBUG("column_stats pointer passed is null");
-    return DEFAULT_SELECTIVITY;
-  }
-
-  size_t numrows = column_stats->GetNumRows();
-
-  // For now only double is supported in stats storage
-  auto top_k = column_stats->GetTopK();
-
-  // Find frequency of the value if present in the top K elements.
-  auto value_frequency_estimate = top_k->EstimateItemCount(value);
-
-  double res = value_frequency_estimate / static_cast<double>(numrows);
-
-  NOISEPAGE_ASSERT(res >= 0 && res <= 1, "res must be in valid range");
-  return res;
-}
 }  // namespace noisepage::optimizer
